Fix incorrect value for auto_accept
We are accepting messages (confirming we've received them and they
should be removed from the UMB queue) manually if we succeed in
scheduling a Celery task to process the message itself. So auto_accept
should be disabled.

Also, add an explicit setting to auto-settle all accepted messages (this
is the default) to indicate the application considers the delivery to be
complete.