Refactor namespace and type attributes of the Component model
This is a mega-commit because it touches so many parts of the code base.
To update the default list of choices, we have to add a new attribute,
which also means an update to all the places that fill out our data
(mainly in the collectors) as well as the API layer.

The list of component types now aligns more closely to the purl spec
(with the exception of the "rpmmod" type that we have to request a new
type for). This means we'll produce purls that can be distributed
outside of Red Hat and potentially reused by other systems that also use
purl for component identification.