Move ordering options to ViewSets and keep core models unordered
This is needed so we don't have to write .order_by() each time
we also want to call .distinct(). It also speeds up some queries
since the database doesn't have to apply ordering when it's not
needed.

For API / view-based queries, these can define a custom ordering
for each individual viewset. Ordering is needed here to prevent
bugs with Django-REST-Framework's pagination logic.