Move tasks which spawn many copies back to the slow queue
This means that the fast queue will contain at most a small number
of tasks, so we can guarantee tasks scheduled there will be run
on time. We need this for our daily monitoring email and daily
load tasks.

Slow tasks spawn many copies of themselves that can fill up a queue.
A large number of tasks in a queue delays other tasks scheduled
to be run there, so these tasks are "slow" in the sense they may
not run right away. "slow" does not refer to how long the task
takes to complete once it starts running, and also does not refer
to whether a task is CPU-bound or I/O-bound.