Remove broken /taxonomy endpoint from ProductModel subclasses
It doesn't work and probably hasn't for a while (ever since we
added channels). There are no links to the /taxonomy endpoints
from any detail view, so users probably aren't even aware that
these pages exist. They also aren't mentioned in our user guide,
although the endpoints are documented in the OpenAPI schema.

Fixing the /taxonomy endpoint is possible, but seems pointless
since the same information is already available on the detail view
for a particular ProductModel subclass.

The /taxonomy endpoint was also never implemented for Channels.
Again, fixing this is possible but duplicates info in the detail view.

The /taxonomy endpoint for Components will stay since it's working.