Make "sources" exact inverse of "provides" to fix an existing bug
"sources" and "provides" should be exact inverses of each other,
so following one, then going back on the other, is always possible.

Currently this is not the case. Consider an SRPM root node with
type SOURCE, a container child node with type PROVIDES, and an RPM
grandchild / leaf node with type PROVIDES.

When saving component taxonomies, the get_provides function looks
for any descendant nodes with type PROVIDES. So the SRPM will have
2 provides for its container child and RPM grandchild.
The container will have 1 provides for its RPM grandchild / leaf
node. The RPM will have 0 provides.

But the get_sources function looks only for root nodes with
parent=None. So the SRPM will have 0 sources. The container will
have one source for its SRPM parent. The RPM will have only 1 source
for its SRPM grandparent, and does not include its container parent
as a source.

This means that following the relationships from the container
to the RPM, then back to the container, is not possible.
It also means that discovering which containers an RPM is embedded
in is not always possible. Containers which are root nodes will
be reported, but not containers which are a child of / provided by
some other node.

To fix this, I changed get_sources to consider all parent nodes
of PROVIDES or PROVIDES_DEV type nodes. Nodes of other
types (REQUIRES, SOURCE) will not be reported by
get_provides, so should not be considered for get_sources.

This is the inverse of get_provides, which looks for all
children with type PROVIDES or PROVIDES_DEV.

We could instead change get_provides, so that it's a no-op when
called on a node with type PROVIDES or PROVIDES_DEV.
In other words, we'd only report provides from SOURCE or REQUIRES
nodes. But this would make discovering RPMs embedded in a container
impossible whenever the container is not a root node, as in the
example above.