Unpin UBI container image's minor version tag, to upgrade to 8.7
This will prevent many CVEs from accumulating in our containers.
I thought UBI 8.6 was still supported, but it looks like the last
UBI container on the 8.6 tag / stream was built 3 months ago.

By pulling the latest image from the ubi8/ubi repo, we'll
automatically get any newer 8.8 / 8.9 minor versions when they're
available. This does not include major version upgrades (and we
don't want those to be automatic), so we'll still need to manually
upgrade to RHEL9 / UBI9 at the end of the RHEL8 / UBI8 lifecycle.