Use only RPM semantics when deciding which component version is latest
If there are two components like bash.el7 and bash.el8, we want to
treat these as different components for different platforms.
We will report the latest version of bash on RHEL 7, and the latest
version of bash on RHEL 8, and allow users to filter which version
of RHEL / which el_match they're interested in (by default, all of them)
This supports layered products like OpenShift, which include different
components depending on which version of RHEL they were built for.