Make latest / root component filters generic and user-controllable
OSC said "too many components" (really versions) were being returned,
so we added a filter to only show the latest version of some component.
This filter was always on when listing "all components for some stream",
and couldn't be turned off by the user.
The latest filter also only worked for root components,
due to relying on SoftwareBuild timestamps.

OSC's requirements have now changed,
and they want to see the latest components in a stream,
including non-root components.
A generic latest filter, and a separate root components filter,
supports the following use cases:

1) All versions of all components in a stream (new default)
2) Latest version of all components in a stream
3) All versions of root components in a stream
4) Latest version of root components in a stream (old default)

Making the filter generic means it now appears in our API documentation,
and it also can be turned on or off by the user,
as needed for a particular query.
This means we shouldn't need to do new dev work,
if requirements change again or some other team has a new usecase.