This commit changes the productstreams attribute of the ProductVariant model from a ForeignKey to a Many-to-Many relationship.
It also consolidates the ofuri generation logic into a single function of ProductModel. This new version uses the ProductNodes to build the ofuri for a Product. This change means that we now need to create the ProductNode tree before creating the ProductModel. Because we need the ProductNode tree in place before generating an ofuri I've also changed the way we build the product[version,stream] attributes so that they are populated when save_product_taxonomy is called.

Another major change is the way we remove a ProductModel from a Component. Now we always reset the product taxonomy from what is in the relations table. Because of the change of productstreams attribute, it was no longer possible to infer sibling in the ProductNode tree before removing parent models.

It also changes the way ProductNodes for the relationship between stream and variant is created in the parse_errata_info function of prod_defs, update_products task so that it always updates the type to DIRECT if an existing INFERRED exists. I do this because quarkus-2.13 at least tried to create both INFERRED and DIRECT relationships. The whole function happens in a transaction, so any switch will take effect only at the end of the update_products function, in an atomic fashion.