Add tox as a dev dependency
tox is the tool used to run the tests both locally and in the CI pipeline, but it is not added as a dependency in any of the requirement files. Therefore, it has to be installed manually and every time the environment is synced using `pip-sync requirements/dev.txt`, it is removed from the environment.

This commit adds tox in the dev.txt requirement file. A couple of packages have also been updated (namely packaging and platformdirs) to avoid conflicts with the dependencies of tox.