Fix links in user guide
The links to the project's OpenAPI specification were broken, as they
were pointing to the corgi repo (now component-registry) at the master
branch (now main). This commit fixes the link to the correct one.

The anchor to the REST API product definitions was also not working, and
now it correctly links to the section.

Since the same information about checking out the OpenAPI specification
to learn about endpoint usage was said twice and very close together,
one of the sentences has been removed to avoid redundancy.